Change up tank right-click logic
When a BlockEntity (e.g. a Furnace) is part of a tank, right-clicking and sneak-right-clicking will now interact with it. You will need the T.I.T. in your main or off-hand to open the tank GUI. Frame blocks without a block entity will still have the GUI opened like previously.